VisibleLayerPersistance.ts
